Accept and decline
The player receiving the referral now has to accept (.refer accept) or decline (.refer decline) before it takes effect. This is to prevent abuse by referring people the user doesn't know.

I haven't tested this on a large scale, or even a smaller stage. I tested it with only 2 accounts, I'm not sure what happens if many people use the feature at exactly the same time.