Build status
Removed the build status because I'm sick of being spammed via email whenever I update something. It's because it can't configure the OS, the code runs as it should without any warnings or errors but I can't stand the emails anymore. I don't think there's anything I can do on my end to fix it so I'm removing it.